fix packaging im my project
