+++ app/src/main/java/com/example/jokerfinder/pojoes/BelongsToCollection.java
@@ -1,4 +1,4 @@
-package com.example.jokerfinder.models;
+package com.example.jokerfinder.pojoes;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
