+++ app/src/main/res/layout/activity_movie_list.xml
@@ -6,7 +6,7 @@
     android:id="@+id/swipeContainer"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.MovieListActivity">
+    tools:context=".features.searchmovie.SearchMovieActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
